GraphQL/requirements.txt
django
graphene-django
django-filter
django-graphql-jwt
django-multiselectfield
Pillow
graphene-file-upload